Add Yeti formula
Signed-Off-By: Max Howell <max@methylblue.com>

The chmod is unrequired, Homebrew does that for you.

Also used full paths to ant as we prefer to be sure about which version of tools we are using.

Fix the yeti script, as write doesn't eval DATA.read by default, and I think
possibly it would be a bad idea, unless we only do it when DATA is passed (not DATA.read as we can't detect that).