pwnat 0.1-beta
pwnat, pronounced "poe-nat", is a tool that allows any number of clients
behind NATs to communicate with a server behind a separate NAT with *no*
port forwarding and *no* DMZ setup on any routers in order to directly
communicate with each other. The server does not need to know anything
about the clients trying to connect.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>