Removed Fortran from NetCDF build
Fortran support has been explicitly disabled in the NetCDF brew pending
the resolution of Issue 72. It is perfectly possible to use the brew to
build a working Fortran NetCDF library if a Fortran compiler is present.
However, configure may fail on 64 bit architectures as Homebrew does not
set the FCFLAGS and FFLAGS environment variables to be compatiable with
those set for CFLAGS.

The best resolution of this issue is formal support for a Fortran
compiler.

A non-existant configure argument related to Szip was removed. This
dependency is satisfied by specifying the location of HDF5.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>