Xapian 1.2.6
Also, changed the way the Python bindings are installed.

The latter change is needed because `/Library/Python` is owned
by root in Lion and would prevent the installation of the compiled
Python module. Instead this is following the technique used in
other formulas -- installing to lib/python and having to add that
directory to `PYTHONPATH`.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>