GraphicsMagick: Build C++ lib by default
Issue #2950 mentions that the C++ lib was disabled because it caused LLVM
builds to fail. GraphicsMagick has since been flagged as `fails_with_llvm` so
this is a moot point and the increase in build time is negligible. A
`--without-magick-plus-plus` option has been added in case people want a lean
build.

Also removed ENV.O3 since it is pointless when `fails_with_llvm` is in effect.