md5sha1sum: fix binary mis-naming issue
The md5sha1sum configure script looks in PREFIX (/usr/local by default)
to see if there are existing md5sum and sha1sum binaries, and adds the
prefix 'ub' to its own binaries if it finds them. This breaks
compatibility with sloccount, which depends on this specific md5
package. Fix it by passing '--prefix=#{prefix}' to configure.

They will override existing md5sum and sha1sum binaries when linked into
the prefix, but some measure of responsibility has to be on the user.

Fixes #5472.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>