bash-completion: fix-up HEAD install
The bash-completion that is installed from head has an entirely new
directory structure. I believe it has some measure of backwards
compatibility in that it will find completion scripts in the old
location, but users still have to source a different file to get it to
actually work, so the caveats have been updated to reflect that.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>