audiofile: Update to 0.3.1
This is an update for audiofile including a new option to enable Code Coverage
support using lcov. There will be a depends_on lcov => :build if the user enables
the new option. I also added a new option to check the build with make check and a
def test function.  There was a compile time option to --enable-valgrind support,
a memory leak testing suite, but homebrewed valgrind fails it's own make check atm.
This update was tested on 10.6.8 x86_64 using 2335, gcc-4.2.1, and clang. It builds
both static and shared libs for your native architecture.

Closes #7635.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>