New Formula: solid
This is a new formula for a C++ library called Solid that is used
for collision detection between objects in 3D space.  This formula
exposes two options to the user, for issues of precision and
rounding error.  This also fixes #5282 and compiles well, using
llvm 2335, gcc-4.2.1, and clang from XCode 4.0.2 on 64bit 10.6.8.
This builds shared and static native libraries.  This is the final
opensource release of this software.

Closes #7776.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>