chromaprint 0.6
This commit updates chromaprint from 0.5 to 0.6. Also, the existing recipe
didn't build examples, which meant that it didn't build fpcalc, which is the
most useful part of the package. Now it builds examples by default, which
introduces a dependency on ffmpeg. If you really want just the libraries,
the option --without-examples will give the old behavior.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>