ffmpeg: use linked_keg when checking optional deps
ffmpeg when built as head checks if x264 is installed using
`Formula.factory.installed?` that won't find x264 stable as it looks for a HEAD
build.  Switch from `.installed?` to `linked_keg.exist?` for all dependencies
as this will find them whether they are installed as head or stable.

Fixes #11996.
Closes #12020.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>