appswitch 1.1.1
Upgrade appswitch to version 1.1.1 which has changes for Lion.
This software has only one 22k source file, but it now needs a
full XCode and xcodebuild to compile it.  Luckily the developer
includes a precompiled binary in the source tarball.  So this
commit just installs that, allowing CLT only users to run it.

Fixes #12198.
Closes #12212.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>